Fixed a compiler error caused by ambiguous call to sqrt

+++ src/numerics/NonlinearSolver.cpp
@@ -1248,7 +1248,7 @@ namespace Cantera {
        * Add junk to the Hessian diagonal
        *  -> Note, testing indicates that this will get too big for ill-conditioned systems.
        */
-      hcol = sqrt(neq_) * 1.0E-7 * hnorm;
+      hcol = sqrt(static_cast<double>(neq_)) * 1.0E-7 * hnorm;
 #ifdef DEBUG_HKM_NOT
       if (hcol > 1.0) {
 	hcol = 1.0E1;
